backend/requirements.txt
Flask==3.1.0
firebase-admin==6.9.0
Flask-Cors==5.0.1
PyJWT==2.10.1
python-dotenv==1.0.1
